pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GraphUniverse"
dynamic = ["version"]
authors = [
    {name = "Louis Van Langendonck", email = "louis.van.langendonck@upc.edu"}
]
readme = "README.md"
description = "Graph Universe: A Library for Generating Inductive Graph Datasets"
license = {file = "LICENSE.txt"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]
requires-python = ">= 3.10"
dependencies=[
    "scikit-optimize",
    "torch",
    "torch-geometric",
    "scipy",
    "torchdiffeq",
    "matplotlib",
    "tqdm",
    "streamlit",
    "omegaconf",
    "seaborn",
    "wandb",
    "tensorboard",
    "einops==0.7.0",
    "tabulate",
    "ipykernel",
    "notebook",
    "jupyterlab",
    "rich",
    "rootutils",
    "scikit-optimize",
]

[project.optional-dependencies]
doc = [
    "jupyter",
    "nbsphinx",
    "nbsphinx_link",
    "numpydoc",
    "sphinx",
    "sphinx_gallery",
    "pydata-sphinx-theme",
    "myst_parser"
]
lint = [
    "pre-commit",
    "ruff>=0.1.0"
]
test = [
    "pytest>=7.0",
    "pytest-cov",
    "coverage",
    "jupyter",
    "mypy",
    "pytest-mock"
]

dev = ["GraphUniverse[test, lint]"]
all = ["GraphUniverse[dev, doc]"]

[project.urls]
repository="https://github.com/LouisVanLangendonck/GraphUniverse"

[tool.setuptools.dynamic]
version = {attr = "graph_universe.__version__"}

[tool.setuptools.packages.find]
include = [
    "graph_universe",
    "graph_universe.*"
]

[tool.coverage.run]
source = ["graph_universe"]
omit = ["*/test/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--import-mode=importlib"

# ============================================================
# Ruff Configuration
# ============================================================

[tool.ruff]
# Maximum line length (matches Black default)
line-length = 100

# Python version for compatibility
target-version = "py310"

# Directories to exclude from linting
exclude = [
    ".git",
    ".pytest_cache",
    "__pycache__",
    "*.egg-info",
    ".venv",
    "venv",
    "build",
    "dist",
]

[tool.ruff.lint]
# Enable specific rule sets
select = [
    "E",     # pycodestyle errors (PEP 8 violations)
    "W",     # pycodestyle warnings
    "F",     # pyflakes (undefined names, unused imports)
    "I",     # isort (import sorting)
    "N",     # pep8-naming (naming conventions)
    "UP",    # pyupgrade (modern Python syntax)
    "B",     # flake8-bugbear (likely bugs)
    "C4",    # flake8-comprehensions (better comprehensions)
    "SIM",   # flake8-simplify (simplification suggestions)
    "NPY",   # numpy-specific rules
    "RUF",   # ruff-specific rules
]

# Rules to ignore
ignore = [
    "E501",   # Line too long (let formatter handle this)
    "B008",   # Function call in default argument (common in ML)
    "N803",   # Argument name should be lowercase (for math: K, P, etc.)
    "N806",   # Variable in function should be lowercase (for math)
    "SIM108", # Use ternary operator (sometimes less readable)
    "NPY002", # Using np.random.Generator instead of np.random.seed and no.random.Uniform etc.
]

# Allow specific violations in specific files
[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",  # Unused import (OK in __init__.py for re-exports)
    "F403",  # Star import (sometimes needed)
]
"test_*.py" = [
    "N802",  # Function name should be lowercase (test_ClassName is OK)
    "N806",  # Variable should be lowercase (OK in tests)
    "B011",  # Assert false (used in tests)
]

# Specific rule configurations
[tool.ruff.lint.isort]
known-first-party = ["graph_universe"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pep8-naming]
# Allow mathematical variable names
extend-ignore-names = ["K", "P", "N", "X", "Y"]

[tool.ruff.format]
# Use double quotes (default)
quote-style = "double"

# Use spaces for indentation
indent-style = "space"

# Respect magic trailing comma (keeps multi-line formatting)
skip-magic-trailing-comma = false
